requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
discord.py==1.6.0
google-api-core==1.26.2
google-auth==1.28.0
google-cloud-core==1.6.0
google-cloud-translate==3.0.2
googleapis-common-protos==1.53.0
grpcio==1.36.1
idna==2.10
iniconfig==1.1.1
libcst==0.3.17
multidict==5.1.0
mypy-extensions==0.4.3
packaging==20.9
pluggy==0.13.1
proto-plus==1.18.1
protobuf==3.15.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.2.2
pytest-asyncio==0.14.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
yarl==1.6.3
